refactor: UserRepository를 interface로 분리
